server/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.1
    # via starlette
click==8.1.3
    # via uvicorn
contourpy==1.0.5
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fastapi==0.85.0
    # via -r requirements.in
fonttools==4.37.3
    # via matplotlib
h11==0.13.0
    # via uvicorn
idna==3.4
    # via anyio
imageio==2.22.0
    # via scikit-image
imutils==0.5.4
    # via -r requirements.in
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.6.0
    # via -r requirements.in
networkx==2.8.6
    # via scikit-image
numpy==1.23.3
    # via
    #   contourpy
    #   imageio
    #   matplotlib
    #   opencv-python
    #   pywavelets
    #   scikit-image
    #   scipy
    #   tifffile
opencv-python==4.6.0.66
    # via -r requirements.in
packaging==21.3
    # via
    #   matplotlib
    #   pytesseract
    #   scikit-image
pillow==9.2.0
    # via
    #   imageio
    #   matplotlib
    #   pytesseract
    #   scikit-image
pydantic==1.10.2
    # via fastapi
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
pytesseract==0.3.10
    # via -r requirements.in
python-dateutil==2.8.2
    # via matplotlib
pywavelets==1.4.1
    # via scikit-image
scikit-image==0.19.3
    # via -r requirements.in
scipy==1.9.1
    # via
    #   -r requirements.in
    #   scikit-image
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.20.4
    # via fastapi
tifffile==2022.8.12
    # via scikit-image
typing-extensions==4.3.0
    # via pydantic
uvicorn==0.18.3
    # via -r requirements.in
